fix: Reset the page on limiter change
Set the page back to page 1 when the user's per page is altered.
Previously if the user's per page is increased and the total number
of pages falls, the user could be outside the normal range of pages
and thus an empty table is rendered.

E.g. Imagine in a table of 100 users total
user is presently on page 4 where 25 users are shown per page
Users per page is now changed to 100 and so there can only be one page
but user is still in page 4 outside the range of pages and thus the
table is shown empty.