feat: Render the user summary cards even on fetch
Previously the user summary cards does not show while the data is being
fetched. The object `userListSummary` in state is first empty {},
then populated when the data is received from the server. This was
changed so that it shows the cards and even a skeleton components
where the values are at.